arrumando numeracao ao lado do icone do carrinho
